Change Vector_t implementation to use s32 instead of long

+++ day01a.c
@@ -5,8 +5,8 @@ int main (int argc, char ** argv) {
     Vector_t v;
     vector_init(&v, 256);
     do {
-        long in;
-        int ret = scanf("%li", &in);
+        int in;
+        int ret = scanf("%i", &in);
         if (ret == 1) {
             vector_push_back(&v, in);
         } else if (ret == EOF) {
@@ -17,10 +17,10 @@ int main (int argc, char ** argv) {
     } while (1);
 
     for (size_t i = 0; i < (v.count-2); ++i) {
-        long target = 2020 - v.data[i];
+        int target = 2020 - v.data[i];
         for (size_t j = i+1; j < v.count; ++j) {
             if (v.data[j] == target) {
-                printf("Answer: %li\n", v.data[i]*v.data[j]);
+                printf("Answer: %i\n", v.data[i]*v.data[j]);
                 exit(EXIT_SUCCESS);
             }
         }
